+++ src/Adapter/AbstractAdapter.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Berlioz\Http\Client\Adapter;
 
 use Berlioz\Http\Client\History\Timings;
+use Berlioz\Http\Client\Stream\InflateStream;
 use Berlioz\Http\Message\Request;
 use Berlioz\Http\Message\Stream;
 use InvalidArgumentException;
@@ -67,36 +68,31 @@ abstract class AbstractAdapter implements AdapterInterface
     /**
      * Create stream from content.
      *
-     * @param string|null $content
+     * @param string|resource|null $content
      * @param array|null $contentEncodingHeader
      *
      * @return StreamInterface
      */
-    protected function createStream(?string $content, ?array $contentEncodingHeader): StreamInterface
+    protected function createStream(mixed $content, ?array $contentEncodingHeader): StreamInterface
     {
-        if (!is_resource($content) && !is_string($content) && !is_null($content)) {
+        if (!$content instanceof StreamInterface &&
+            !is_resource($content) &&
+            !is_string($content) &&
+            !is_null($content)) {
             throw new InvalidArgumentException('Parameter must be a resource type, string or null value.');
         }
 
-        $stream = new Stream();
-
-        if (empty($content)) {
-            return $stream;
+        $stream = $content instanceof StreamInterface ? $content : new Stream(is_resource($content) ? $content : null);
+        if (is_string($content) && !empty($content)) {
+            $stream->write($content);
         }
 
         if (!empty($contentEncodingHeader)) {
             // Gzip
-            if (in_array('gzip', $contentEncodingHeader)) {
-                $content = gzdecode($content);
-            }
-
-            // Deflate
-            if (in_array('deflate', $contentEncodingHeader)) {
-                $content = gzinflate($content);
+            if (in_array('gzip', $contentEncodingHeader) || in_array('deflate', $contentEncodingHeader)) {
+                return new InflateStream($stream);
             }
         }
-
-        $stream->write($content);
 
         return $stream;
     }