+++ tests/server/request.php
@@ -40,6 +40,6 @@ if ($_GET['encoding'] ?? null) {
 
 print match ($_GET['encoding'] ?? null) {
     'gzip' => gzencode($contents),
-    'deflate' => gzdeflate($contents),
+    'deflate' => gzdeflate($contents, encoding: ZLIB_ENCODING_DEFLATE),
     default => $contents,
 };